+++ lib/src/televerse/models/message_or_boolean.dart
@@ -1,5 +1,12 @@
 part of televerse.models;
 
+/// **Message or Boolean**
+///
+/// This class is used to represent a message or a boolean.
+/// There are some methods in the Telegram API that can return either a message or a boolean. To represent this, we use this class.
+/// For example [Televerse.editMessageLiveLocation], [Televerse.stopMessageLiveLocation] are methods that can return either a message or a boolean depending on the parameters you pass.
+///
+/// This class is used to represent this.
 class MessageOrBoolean {
   Message? message;
   bool? boolean;
@@ -17,6 +24,13 @@ class MessageOrBoolean {
     }
   }
 
+  /// **Type**
+  ///
+  /// This getter returns the type of the [MessageOrBoolean] class.
+  ///
+  /// This can either be [MsgOrBool.boolean] or [MsgOrBool.message].
+  ///
+  /// This can be used to identify either [bool] or [Message] is returned from Telegram API.
   MsgOrBool get type {
     if (message != null) {
       return MsgOrBool.message;
@@ -26,6 +40,10 @@ class MessageOrBoolean {
   }
 }
 
+/// **Message or Boolean**
+///
+/// This enum is used to identify whether the [MessageOrBoolean] class contains a message or a boolean.
+/// The [MessageOrBoolean.type] getter returns this enum.
 enum MsgOrBool {
   boolean,
   message,