+++ lib/src/televerse/televerse.dart
@@ -1605,9 +1605,9 @@ class Televerse extends Event {
   /// Use this method to add a message to the list of pinned messages in a chat. If the chat is not a private chat, the bot must be an administrator in the chat for this to work and must have the 'can_pin_messages' administrator right in a supergroup or 'can_edit_messages' administrator right in a channel. Returns True on success.
   Future<bool> pinChatMessage(
     ID chatId,
-    int messageId,
+    int messageId, {
     bool? disableNotification,
-  ) async {
+  }) async {
     Map<String, dynamic> params = {
       "chat_id": chatId.id,
       "message_id": messageId,