+++ lib/televerse.dart
@@ -22,6 +22,7 @@ part 'src/utils/http.dart';
 part 'src/utils/utils.dart';
 part 'src/televerse/filters/filters.dart';
 part 'src/televerse/context/mixins/message_mixin.dart';
+part 'src/televerse/context/mixins/management.dart';
 
 /// The main class of the library.
 ///